+++ client/test/fixture/index.jsx
@@ -3,6 +3,8 @@ import './css/test.scss'
 
 class App {
 	render() {
+		$('.class10')
+
 		return (
 			<div>
 				<div id="id1" />
@@ -13,6 +15,7 @@ class App {
 				<div className={`any-other-class ` + any_variable_or_expression + ` class5`} />
 				<div className={style.class6} />
 				<div className={style['class7']} />
+				<div class={['class8', { class9: this.show }, 'class10']}></div>
 			</div>
 		)
 	}