+++ server/src/languages/html/html-scanner.ts
@@ -1,6 +1,9 @@
+import {resolveImportPath} from '../../helpers/file'
 import {firstMatch} from '../../helpers/utils'
 import {SimpleSelector} from '../common/simple-selector'
 import {TextScanner} from '../common/text-scanner'
+import * as path from 'path'
+import {URI} from 'vscode-uri'
 
 
 /*
@@ -28,73 +31,57 @@ export class HTMLScanner extends TextScanner {
 
 	/** Scan a HTML document from a specified offset to find a CSS selector. */
 	scanForSelector(): SimpleSelector | null {
-		let word = this.readLeftWord()
-		if (!word) {
-			return null
-		}
-		
-		let char = this.peekLeftChar()
-		let leftOffset = this.offset + 1
-
-		if (char === '<') {
-			return SimpleSelector.create(word, leftOffset)
+		// <tag...>
+		let match = this.match(/<([\w-]+)/g)
+		if (match) {
+			return SimpleSelector.create(match.text, match.index)
 		}
 
-		this.readLeftUntil(['<', '\'', '"'])
-		if (this.peekRightChar(1) === '<') {
-			return null
-		}
+		// <tag
+		// 	 id="a'
+		// 	 class="a"
+		// 	 class="a b"
+		// >
+		match = this.match(
+			/<[\w-]+\s*(.*?)>/g,
+			/(?<type>id|class)\s*=\s*['"](.*?)['"]/g,
+			/([\w-]+)/g,
+		)
 
-		/*
-		may be in left:
-			class="a
-			id="a'
-			class="a b
-			class="a" b
-		have a very low possibility to meet '<tag a="class=" b', ignore it.
-		the good part is it can get selectors in any place, no matter the code format.
-		*/
-		if (this.peekLeftChar() === '\\') {
-			this.moveLeft()
-		}
-
-		this.skipLeftWhiteSpaces()
-		if (this.readLeftChar() !== '=') {
-			return null
-		}
-
-		this.skipLeftWhiteSpaces()
-		let attribute = this.readLeftWord().toLowerCase()
-
-		if (attribute === 'class' || attribute === 'id') {
-			let raw = (attribute === 'class' ? '.' : '#') + word
-			return SimpleSelector.create(raw, leftOffset)
+		if (match) {
+			if (match.groups.type === 'id') {
+				return SimpleSelector.create('#' + match.text, match.index)
+			}
+			else if (match.groups.type === 'class') {
+				return SimpleSelector.create('.' + match.text, match.index)
+			}
 		}
 
 		return null
 	}
 
 	/** Scan for relative import path. */
-	scanForImportPath() {
-		this.readLeftUntil(['<', '>'])
-		if (this.peekRightChar(1) !== '<') {
-			return null
+	async scanForImportPath() {
+		let match = this.match(/<(?<tag>[\w-]+)(\s*.*?)>/g)
+		let importPath: string | null = null
+
+		if (match) {
+			let tag = match.groups.tag
+			let linkStyleRE = /\brel\s*=\s*['"]stylesheet['"]/
+			let srcRE = /\bsrc\s*=['"](.*?)['"]/
+			let hrefRE = /\bhref\s*=['"](.*?)['"]/
+
+			if (tag === 'link' && linkStyleRE.test(match.text)) {
+				importPath = firstMatch(match.text, hrefRE)
+			}
+
+			if (tag === 'style') {
+				importPath = firstMatch(match.text, srcRE)
+			}
 		}
 
-		this.moveRight()
-
-		let code = this.readRightUntil(['>'])
-		let tag = firstMatch(code, /^<(\w+)/)
-		let linkRE = /<link[^>]+rel\s*=\s*['"]stylesheet['"]/
-		let hrefRE = /\bhref\s*=['"](.*?)['"]/
-		let styleRE = /<style[^>]+src\s*=['"](.*?)['"]/
-
-		if (tag === 'link' && linkRE.test(code)) {
-			return firstMatch(code, hrefRE)
-		}
-
-		if (tag === 'style') {
-			return firstMatch(code, styleRE)
+		if (importPath) {
+			return await resolveImportPath(path.dirname(URI.parse(this.document.uri).fsPath), importPath)
 		}
 
 		return null