+++ server/src/languages/jsx/jsx-scanner.ts
@@ -2,7 +2,7 @@ import {SimpleSelector} from '../common/simple-selector'
 import {TextScanner} from '../common/text-scanner'
 import * as path from 'path'
 import {URI} from 'vscode-uri'
-import * as fs from 'fs-extra'
+import {getPathExtension, resolveImportPath} from '../../helpers/file'
 
 
 /** 
@@ -11,118 +11,127 @@ import * as fs from 'fs-extra'
  */
 export class JSXScanner extends TextScanner {
 
+	private supportedLanguages = ['css', 'less', 'scss']
+
 	/** Scan a JSX / JS / TS / TSX document from a specified offset to find a CSS selector. */
 	async scanSelector(): Promise<SimpleSelector | null> {
-		let inExpression = false
+		// <tag
+		// 	 id="a'
+		// 	 class="a"
+		// 	 class="a b"
+		// >
+		let match = this.match(
+			/<\w+\s*(.*?)>/g,
+			/\b(?<type>id|class|className)\s*=\s*['"`](.*?)['"`]/g,
+			/([\w-]+)/g,
+		)
 
-		let attributeValue = this.readLeftWord()
-		let wordLeftOffset = this.offset + 1
-
-		if (!attributeValue) {
-			return null
-		}
-		
-		
-		// `.xxx`
-		if (this.peekLeftChar() === '.') {
-			this.readLeftChar()
-
-			this.skipLeftWhiteSpaces()
-			let attributeName = this.readLeftWord()
-
-			// For Flit syntax `:class.property=...`
-			if (attributeName === 'class') {
-				let raw = '.' + attributeValue
-				return SimpleSelector.create(raw, wordLeftOffset)
+		if (match) {
+			if (match.groups.type === 'id') {
+				return SimpleSelector.create('#' + match.text, match.index)
 			}
+			else if (match.groups.type === 'class' || match.groups.type === 'className') {
+				return SimpleSelector.create('.' + match.text, match.index)
+			}
+		}
 
-			// Module CSS, e.g. `className={style.className}`.
+
+		// Syntax `:class.property=...`
+		match = this.match(/\bclass\.([\w-]+)/g)
+		if (match) {
+			return SimpleSelector.create('.' + match.text, match.index)
+		}
+
+
+		// Syntax: `:class=${{property: boolean}}`.
+		match = this.match(
+			/\bclass\s*=\s*\$\{\s*\{(.*?)\}\s*\}/g,
+			/(\w+)\s*:/g,
+		)
+		if (match) {
+			return SimpleSelector.create('.' + match.text, match.index)
+		}
+
+
+		// React syntax:
+		// `class={['...']}, '...' part
+		// `class={'...'}
+		match = this.match(
+			/\b(?:class|className)\s*=\s*\{((?:\{[\s\S]*?\}|.)*?)\}/g,
+			/['"`](.*?)['"`]/g,
+			/([\w-]+)/g,
+		)
+		if (match) {
+			return SimpleSelector.create('.' + match.text, match.index)
+		}
+
+
+		// React syntax:
+		// `class={[..., {...}]}, {...} part.
+		match = this.match(
+			/\b(?:class|className)\s*=\s*\{((?:\{[\s\S]*?\}|.)*?)\}/g,
+			/\{(.*?)\}/g,
+			/(\w+)\s*:/g,
+		)
+		if (match) {
+			return SimpleSelector.create('.' + match.text, match.index)
+		}
+
+
+		// Due to https://github.com/gajus/babel-plugin-react-css-modules and issue #60.
+		// `styleName='...'.
+		match = this.match(
+			/\bstyleName\s*=\s*['"`](.*?)['"`]/g,
+			/([\w-]+)/g,
+		)
+		if (match) {
+			return this.scanDefaultCSSModule(match.text, match.index)
+		}
+
+
+		// React Module CSS, e.g.
+		// `class={style.className}`.
+		// `class={style['class-name']}`.
+		match = this.match(
+			/\b(?:class|className)\s*=\s*\{(.*?)\}/g,
+			/(?<moduleName>\w+)(?:\.(\w+)|\[\s*['"`](\w+)['"`]\s*\])/,
+		)
+		if (match) {
+			return this.scanCSSModule(match.groups.moduelName, match.text, match.index)
+		}
+
+
+		// jQuery selector, e.g.
+		// `$('.abc')`
+		match = this.match(
+			/\$\((.*?)\)/g,
+			/['"`](.*?)['"`]/g,
+			/(?<identifier>^|\s|.|#)([\w-]+)/g,
+		)
+		if (match) {
+			if (match.groups.identifier === '#' || match.groups.identifier === '.') {
+				return SimpleSelector.create(match.groups.identifier + match.text, match.index)
+			}
 			else {
-				return this.scanCSSModule(attributeValue, wordLeftOffset)
+				return SimpleSelector.create(match.text, match.index)
 			}
 		}
 
-
-		// Module CSS, e.g. `className={style['class-name']}`.
-		if ((this.peekLeftChar() === '"' || this.peekLeftChar() === '\'') && this.peekLeftCharSkipWhiteSpaces(1) === '[') {
-			this.readLeftUntil(['['])
-			return this.scanCSSModule(attributeValue, wordLeftOffset)
-		}
-
-		this.readLeftUntil(['<', '\'', '"', '`', '{'])
-
-		// Compare with `html-scanner`, here should ignore `<tagName>`.
-		if (this.peekRightChar(1) === '<') {
-			return null
-		}
-
-
-		// Skip expression left boundary `{`.
-		this.skipLeftWhiteSpaces()
-		if (this.peekLeftChar() !== '=') {
-
-			// Assume it's in `className={...[HERE]...}` or `class="..."`
-			this.readLeftUntil(['<', '{', '}'])
-			if (this.peekRightChar(1) !== '{') {
-				return null
-			}
-
-			// Flit syntax `:class=${{property: boolean}}`.
-			if (this.peekLeftCharSkipWhiteSpaces() === '{' && this.peekLeftCharSkipWhiteSpaces(1) === '$') {
-				this.readLeftUntil(['$'])
-			}
-
-			inExpression = true
-		}
-
-
-		// Read `=`.
-		this.skipLeftWhiteSpaces()
-		if (this.readLeftChar() !== '=') {
-			return null
-		}
-		
-		this.skipLeftWhiteSpaces()
-		let attributeName = this.readLeftWord()
-
-		if (attributeName === 'className' || attributeName === 'class' || attributeName === 'id' && !inExpression) {
-			let raw = (attributeName === 'id' ? '#' : '.') + attributeValue
-			return SimpleSelector.create(raw, wordLeftOffset)
-		}
 
 		return null
 	}
 
 	/** Scan imported CSS module. */
-	private async scanCSSModule(attributeValue: string, wordLeftOffset: number): Promise<SimpleSelector | null> {
-		let moduleVariable = this.readLeftWord()
-		if (!moduleVariable) {
-			return null
-		}
-
-		this.readLeftUntil(['{'])
-		this.skipLeftWhiteSpaces()
-
-		if (this.readLeftChar() !== '=') {
-			return null
-		}
-
-		// Must be `className={style.className}`, or it will popup frequently even type `a.b`.
-		this.skipLeftWhiteSpaces()
-		let className = this.readLeftWord()
-		if (className !== 'class' && className !== 'className') {
-			return null
-		}
-
-		let modulePath = this.parseImportedPathFromVariableName(moduleVariable)
+	private async scanCSSModule(moduleName: string, moduleProperty: string, wordLeftOffset: number): Promise<SimpleSelector | null> {
+		let modulePath = this.parseImportedPathFromVariableName(moduleName)
 		if (modulePath) {
-			let fullPath = path.resolve(path.dirname(URI.parse(this.document.uri).fsPath), modulePath)
-			if (await fs.pathExists(fullPath)) {
-				return SimpleSelector.create('.' + attributeValue, wordLeftOffset, URI.file(fullPath).toString())
+			let fullPath = await resolveImportPath(path.dirname(URI.parse(this.document.uri).fsPath), modulePath)
+			if (fullPath) {
+				return SimpleSelector.create('.' + moduleProperty, wordLeftOffset, URI.file(fullPath).toString())
 			}
 		}
 
-		return SimpleSelector.create('.' + attributeValue, wordLeftOffset)
+		return SimpleSelector.create('.' + moduleProperty, wordLeftOffset)
 	}
 
 	/** Parse `import ...`. */
@@ -140,26 +149,45 @@ export class JSXScanner extends TextScanner {
 		return null
 	}
 
+	/** Scan imported CSS module. */
+	private async scanDefaultCSSModule(moduleProperty: string, wordLeftOffset: number): Promise<SimpleSelector | null> {
+		let modulePath = this.parseDefaultImportedPath()
+		if (modulePath) {
+			let fullPath = await resolveImportPath(path.dirname(URI.parse(this.document.uri).fsPath), modulePath)
+			if (fullPath) {
+				return SimpleSelector.create('.' + moduleProperty, wordLeftOffset, URI.file(fullPath).toString())
+			}
+		}
+
+		return SimpleSelector.create('.' + moduleProperty, wordLeftOffset)
+	}
+
+	/** Parse `import '....css'`. */
+	private parseDefaultImportedPath(): string | null {
+		let re = /import\s+['"`](.+?)['"`]/g
+		let match: RegExpExecArray | null
+
+		while (match = re.exec(this.text)) {
+			let path = match[1]
+			let extension = getPathExtension(path)
+			if (this.supportedLanguages.includes(extension)) {
+				return path
+			}
+		}
+
+		return null
+	}
+
 	/** Scan for relative import path. */
-	scanForImportPath() {
-		this.peekLeftChar
+	async scanForImportPath() {
 
 		// import * from '...'
 		// import abc from '...'
 		// import '...'
-		let re = /import\s+(?:(?:\w+|\*)\s+from\s+)?['"`](.+?)['"`]/g
-		let match: RegExpExecArray | null
 
-		re.lastIndex = this.offset - 1024
-
-		while (match = re.exec(this.text)) {
-			// |'...'|, `|` marks location of start index and end index.
-			let endIndex = re.lastIndex
-			let startIndex = re.lastIndex - match[1].length - 2
-
-			if (startIndex <= this.offset && endIndex > this.offset) {
-				return match[1]
-			}
+		let modulePath = this.match(/import\s+(?:(?:\w+|\*)\s+from\s+)?['"`](.+?)['"`]/g)?.text || null
+		if (modulePath) {
+			return await resolveImportPath(path.dirname(URI.parse(this.document.uri).fsPath), modulePath)
 		}
 
 		return null